requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
numpy==2.1.3
gunicorn==22.0.0
dj-database-url==2.2.0
whitenoise==6.7.0
django-cors-headers==4.4.0
psycopg[binary]==3.2.9
drf-spectacular==0.27.2
